+++ library/core/src/main/java/com/google/android/exoplayer2/offline/DownloadHelper.java
@@ -15,6 +15,7 @@
  */
 package com.google.android.exoplayer2.offline;
 
+import android.content.Context;
 import android.net.Uri;
 import android.os.Handler;
 import android.os.HandlerThread;
@@ -82,12 +83,25 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
  */
 public final class DownloadHelper {
 
+  /** Default track selection parameters for downloading, but without any viewport constraints. */
+  public static final Parameters DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT =
+      Parameters.DEFAULT_WITHOUT_VIEWPORT.buildUpon().setForceHighestSupportedBitrate(true).build();
+
   /**
-   * The default parameters used for track selection for downloading. This default selects the
-   * highest bitrate audio and video tracks which are supported by the renderers.
+   * @deprecated This instance does not have viewport constraints configured for the primary
+   *     display. Use {@link #getDefaultTrackSelectorParameters(Context)} instead.
    */
+  @Deprecated
   public static final DefaultTrackSelector.Parameters DEFAULT_TRACK_SELECTOR_PARAMETERS =
-      new DefaultTrackSelector.ParametersBuilder().setForceHighestSupportedBitrate(true).build();
+      DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT;
+
+  /** Returns the default parameters used for track selection for downloading. */
+  public static DefaultTrackSelector.Parameters getDefaultTrackSelectorParameters(Context context) {
+    return Parameters.getDefaults(context)
+        .buildUpon()
+        .setForceHighestSupportedBitrate(true)
+        .build();
+  }
 
   /** A callback to be notified when the {@link DownloadHelper} is prepared. */
   public interface Callback {
@@ -120,12 +134,9 @@ public final class DownloadHelper {
   private static final Constructor<? extends MediaSourceFactory> HLS_FACTORY_CONSTRUCTOR =
       getConstructor("com.google.android.exoplayer2.source.hls.HlsMediaSource$Factory");
 
-  /**
-   * Creates a {@link DownloadHelper} for progressive streams.
-   *
-   * @param uri A stream {@link Uri}.
-   * @return A {@link DownloadHelper} for progressive streams.
-   */
+  /** @deprecated Use {@link #forProgressive(Context, Uri)} */
+  @Deprecated
+  @SuppressWarnings("deprecation")
   public static DownloadHelper forProgressive(Uri uri) {
     return forProgressive(uri, /* cacheKey= */ null);
   }
@@ -133,23 +144,60 @@ public final class DownloadHelper {
   /**
    * Creates a {@link DownloadHelper} for progressive streams.
    *
+   * @param context Any {@link Context}.
    * @param uri A stream {@link Uri}.
-   * @param cacheKey An optional cache key.
    * @return A {@link DownloadHelper} for progressive streams.
    */
+  public static DownloadHelper forProgressive(Context context, Uri uri) {
+    return forProgressive(context, uri, /* cacheKey= */ null);
+  }
+
+  /** @deprecated Use {@link #forProgressive(Context, Uri, String)} */
+  @Deprecated
   public static DownloadHelper forProgressive(Uri uri, @Nullable String cacheKey) {
     return new DownloadHelper(
         DownloadRequest.TYPE_PROGRESSIVE,
         uri,
         cacheKey,
         /* mediaSource= */ null,
-        DEFAULT_TRACK_SELECTOR_PARAMETERS,
+        DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT,
         /* rendererCapabilities= */ new RendererCapabilities[0]);
+  }
+
+  /**
+   * Creates a {@link DownloadHelper} for progressive streams.
+   *
+   * @param context Any {@link Context}.
+   * @param uri A stream {@link Uri}.
+   * @param cacheKey An optional cache key.
+   * @return A {@link DownloadHelper} for progressive streams.
+   */
+  public static DownloadHelper forProgressive(Context context, Uri uri, @Nullable String cacheKey) {
+    return new DownloadHelper(
+        DownloadRequest.TYPE_PROGRESSIVE,
+        uri,
+        cacheKey,
+        /* mediaSource= */ null,
+        getDefaultTrackSelectorParameters(context),
+        /* rendererCapabilities= */ new RendererCapabilities[0]);
+  }
+
+  /** @deprecated Use {@link #forDash(Context, Uri, Factory, RenderersFactory)} */
+  @Deprecated
+  public static DownloadHelper forDash(
+      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+    return forDash(
+        uri,
+        dataSourceFactory,
+        renderersFactory,
+        /* drmSessionManager= */ null,
+        DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT);
   }
 
   /**
    * Creates a {@link DownloadHelper} for DASH streams.
    *
+   * @param context Any {@link Context}.
    * @param uri A manifest {@link Uri}.
    * @param dataSourceFactory A {@link DataSource.Factory} used to load the manifest.
    * @param renderersFactory A {@link RenderersFactory} creating the renderers for which tracks are
@@ -158,13 +206,16 @@ public final class DownloadHelper {
    * @throws IllegalStateException If the DASH module is missing.
    */
   public static DownloadHelper forDash(
-      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+      Context context,
+      Uri uri,
+      DataSource.Factory dataSourceFactory,
+      RenderersFactory renderersFactory) {
     return forDash(
         uri,
         dataSourceFactory,
         renderersFactory,
         /* drmSessionManager= */ null,
-        DEFAULT_TRACK_SELECTOR_PARAMETERS);
+        getDefaultTrackSelectorParameters(context));
   }
 
   /**
@@ -197,9 +248,22 @@ public final class DownloadHelper {
         Util.getRendererCapabilities(renderersFactory, drmSessionManager));
   }
 
+  /** @deprecated Use {@link #forHls(Context, Uri, Factory, RenderersFactory)} */
+  @Deprecated
+  public static DownloadHelper forHls(
+      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+    return forHls(
+        uri,
+        dataSourceFactory,
+        renderersFactory,
+        /* drmSessionManager= */ null,
+        DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT);
+  }
+
   /**
    * Creates a {@link DownloadHelper} for HLS streams.
    *
+   * @param context Any {@link Context}.
    * @param uri A playlist {@link Uri}.
    * @param dataSourceFactory A {@link DataSource.Factory} used to load the playlist.
    * @param renderersFactory A {@link RenderersFactory} creating the renderers for which tracks are
@@ -208,13 +272,16 @@ public final class DownloadHelper {
    * @throws IllegalStateException If the HLS module is missing.
    */
   public static DownloadHelper forHls(
-      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+      Context context,
+      Uri uri,
+      DataSource.Factory dataSourceFactory,
+      RenderersFactory renderersFactory) {
     return forHls(
         uri,
         dataSourceFactory,
         renderersFactory,
         /* drmSessionManager= */ null,
-        DEFAULT_TRACK_SELECTOR_PARAMETERS);
+        getDefaultTrackSelectorParameters(context));
   }
 
   /**
@@ -247,9 +314,22 @@ public final class DownloadHelper {
         Util.getRendererCapabilities(renderersFactory, drmSessionManager));
   }
 
+  /** @deprecated Use {@link #forSmoothStreaming(Context, Uri, Factory, RenderersFactory)} */
+  @Deprecated
+  public static DownloadHelper forSmoothStreaming(
+      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+    return forSmoothStreaming(
+        uri,
+        dataSourceFactory,
+        renderersFactory,
+        /* drmSessionManager= */ null,
+        DEFAULT_TRACK_SELECTOR_PARAMETERS_WITHOUT_VIEWPORT);
+  }
+
   /**
    * Creates a {@link DownloadHelper} for SmoothStreaming streams.
    *
+   * @param context Any {@link Context}.
    * @param uri A manifest {@link Uri}.
    * @param dataSourceFactory A {@link DataSource.Factory} used to load the manifest.
    * @param renderersFactory A {@link RenderersFactory} creating the renderers for which tracks are
@@ -258,13 +338,16 @@ public final class DownloadHelper {
    * @throws IllegalStateException If the SmoothStreaming module is missing.
    */
   public static DownloadHelper forSmoothStreaming(
-      Uri uri, DataSource.Factory dataSourceFactory, RenderersFactory renderersFactory) {
+      Context context,
+      Uri uri,
+      DataSource.Factory dataSourceFactory,
+      RenderersFactory renderersFactory) {
     return forSmoothStreaming(
         uri,
         dataSourceFactory,
         renderersFactory,
         /* drmSessionManager= */ null,
-        DEFAULT_TRACK_SELECTOR_PARAMETERS);
+        getDefaultTrackSelectorParameters(context));
   }
 
   /**
@@ -370,10 +453,10 @@ public final class DownloadHelper {
     this.uri = uri;
     this.cacheKey = cacheKey;
     this.mediaSource = mediaSource;
-    this.trackSelector = new DefaultTrackSelector(new DownloadTrackSelection.Factory());
+    this.trackSelector =
+        new DefaultTrackSelector(trackSelectorParameters, new DownloadTrackSelection.Factory());
     this.rendererCapabilities = rendererCapabilities;
     this.scratchSet = new SparseIntArray();
-    trackSelector.setParameters(trackSelectorParameters);
     trackSelector.init(/* listener= */ () -> {}, new DummyBandwidthMeter());
     callbackHandler = new Handler(Util.getLooper());
     window = new Timeline.Window();